backend/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
transformers
torch
